src/requirements.txt
# requirements.txt
"""
Python dependencies for QuadFusion fraud detection system.
Complete package list for mobile-optimized fraud detection with pretrained model support.
"""

# Core ML libraries
numpy>=1.21.0,<1.25.0
scikit-learn>=1.2.0,<1.4.0
scipy>=1.9.0,<1.12.0

# Deep learning frameworks - mobile optimized versions
tensorflow>=2.13.0,<2.16.0
tensorflow-lite>=2.13.0
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
torchaudio>=2.0.0,<2.2.0

# Model conversion and optimization
onnx>=1.14.0,<1.16.0
onnxruntime>=1.15.0,<1.17.0
onnxruntime-tools>=1.7.0
tensorflow-model-optimization>=0.7.2,<0.8.0
torch-tensorrt>=1.4.0

# Pretrained model repositories
transformers>=4.30.0,<4.40.0
huggingface-hub>=0.15.0,<0.20.0
timm>=0.9.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0

# Audio processing and pretrained models
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0,<0.13.0
speechbrain>=0.5.15,<1.0.0
pyannote-audio>=3.1.0,<4.0.0
whisper-openai>=20231117

# Computer vision and pretrained models
opencv-python>=4.8.0,<5.0.0
opencv-contrib-python>=4.8.0,<5.0.0
Pillow>=9.5.0,<11.0.0
face-recognition>=1.3.0,<2.0.0
mediapipe>=0.10.0,<0.11.0
ultralytics>=8.0.0,<9.0.0

# Data handling and processing
pandas>=2.0.0,<2.2.0
h5py>=3.8.0,<3.11.0
pyarrow>=12.0.0,<15.0.0
polars>=0.18.0,<1.0.0

# Cryptography and security
cryptography>=41.0.0,<42.0.0
pycryptodome>=3.18.0,<4.0.0
keyring>=24.0.0,<25.0.0

# Configuration and environment management
PyYAML>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0
hydra-core>=1.3.0,<2.0.0
omegaconf>=2.3.0,<3.0.0

# Mobile deployment and edge computing
coremltools>=7.0,<8.0
openvino>=2023.0.0,<2024.0.0
apache-tvm>=0.13.0,<1.0.0

# Performance monitoring and optimization
psutil>=5.9.0,<6.0.0
memory-profiler>=0.61.0,<1.0.0
py-spy>=0.3.14,<1.0.0
nvidia-ml-py>=12.0.0

# Utilities and helpers
tqdm>=4.65.0,<5.0.0
joblib>=1.3.0,<2.0.0
cloudpickle>=2.2.0,<3.0.0
dill>=0.3.6,<1.0.0
click>=8.1.0,<9.0.0
rich>=13.0.0,<14.0.0

# Async and concurrent processing
asyncio-mqtt>=0.16.0,<1.0.0
aiofiles>=23.0.0,<24.0.0
uvloop>=0.17.0,<1.0.0

# Testing and quality assurance
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-benchmark>=4.0.0,<5.0.0

# Code quality and formatting
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
pre-commit>=3.3.0,<4.0.0

# Documentation
sphinx>=7.1.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0

# Jupyter and interactive development
jupyter>=1.0.0,<2.0.0
ipywidgets>=8.0.0,<9.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0

# API and web integration
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0,<1.0.0
requests>=2.31.0,<3.0.0
websockets>=11.0.0,<12.0.0

# Database connectivity
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.11.0,<2.0.0
sqlite3

# Additional mobile optimization tools
quantization-toolkit>=1.0.0
neural-compressor>=2.2.0
intel-extension-for-pytorch>=2.0.0

# Platform-specific optimizations (conditional installs)
# For CUDA support (if available)
# nvidia-cublas-cu11>=11.10.3.66
# nvidia-cudnn-cu11>=8.5.0.96

# For Apple Silicon optimization
# tensorflow-macos>=2.13.0; platform_machine=="arm64" and platform_system=="Darwin"
# tensorflow-metal>=1.0.0; platform_machine=="arm64" and platform_system=="Darwin"
